✨Feat: 댓글 CRUD 구현
댓글 작성

Resolves: #11

+++ src/main/kotlin/com/teamsparta/buysell/domain/exception/ModelNotFoundException.kt
@@ -1,4 +1,4 @@
 package com.teamsparta.buysell.domain.exception
 
-data class ModelNotFoundException(val modelName: String, val id: Long?) :
+data class ModelNotFoundException(val modelName: String, val id: Int?) :
     RuntimeException("Model $modelName not found with given id: $id")
